Added comments and minor fix
