--- erpnext/stock/report/eway_bill/eway_bill.json
@@ -1,32 +1,0 @@
-{
- "add_total_row": 0, 
- "apply_user_permissions": 1, 
- "creation": "2018-07-13 19:59:18.922829", 
- "disabled": 0, 
- "docstatus": 0, 
- "doctype": "Report", 
- "idx": 0, 
- "is_standard": "Yes", 
- "modified": "2018-07-13 19:59:18.922829", 
- "modified_by": "Administrator", 
- "module": "Stock", 
- "name": "Eway Bill", 
- "owner": "Administrator", 
- "ref_doctype": "Delivery Note", 
- "report_name": "Eway Bill", 
- "report_type": "Script Report", 
- "roles": [
-  {
-   "role": "Stock User"
-  }, 
-  {
-   "role": "Stock Manager"
-  }, 
-  {
-   "role": "Sales User"
-  }, 
-  {
-   "role": "Accounts User"
-  }
- ]
-}
